+++ crates/engine/body-ir/src/resolution/infer/mod.rs
@@ -3,9 +3,11 @@
 //! The persisted Body IR model stores ordinary `Ty` facts. This module only maps body expression
 //! and binding slots to the transient inference table owned by `rg_ty`.
 
+mod call;
 mod context;
 mod type_ref;
 
+pub(super) use call::BodyCallInference;
 pub(super) use context::BodyInferenceCtx;
 pub(super) use type_ref::{InferTypeRefProjector, InferTypeSubst};
